Updated Go version in Docker and workflows
The Go version has been updated in the Dockerfile as well as in the pull-request and release workflows. The previous version was 1.21.x, which has now been upgraded to 1.22.x. This update ensures that we are using the latest stable release of Go for our development, testing, and deployment processes.