Update go version (#49)
* Update go version

* Updated Go version in Docker and workflows

The Go version has been updated in the Dockerfile as well as in the pull-request and release workflows. The previous version was 1.21.x, which has now been upgraded to 1.22.x. This update ensures that we are using the latest stable release of Go for our development, testing, and deployment processes.

* Update Node.js version to 20.x

The Node.js version has been updated from 16.x to 20.x across multiple files. This includes changes in the Dockerfile, GitHub workflows (pull-request.yml and release.yml), and package.json in the dns-record example.